Changes at 'correcting comments' feature
Now the 'forum_correction.xhtml' snippet doesn't present the 'summernote'
textarea which allowed users to write his corrections. Instead of that,
now it shows a list of inputTexts. The comment to correct is split into
several sentences, by ' .' or '\n' separators. So each of inputTexts
shows a different sentence, to allow users to edit them separately.

The fact of splitting sentences and correcting them separately will help to avoid
errors at the syntactic analysis, which will be neccesary to the
'learning' feature.