Add Feedback functionality and JUnit tests
* Feedback feature:
there was added a formulary in the forum page, which allows to introduce
one sentence and get a possible grammar correction for it.
This feature is based on the stored sentences, created when users publish
corrections in the forum: that corrected sentences will be suggested to
users
with the same doubts.
This is how application 'learns' through participations of users,
and then use it to help them.

* Project "loqua-aggreggator":
the 'tests' package have been added to implement unit
tests against the aggregator class 'Parser.java' (which reads the content
from
xml downloaded files of RSS feeds, and converts it into 'Threads' in the
forum).

* Project "loqua-test-unit":
contains the unit tests for business/persistence layer methods.

* Project "loqua-test-syntaxis":
contains unit tests only dedicated to the task of suggesting grammar
corrections.